Significantly expanded the constitution. At this point,we've got basically everything required by the template.·
I also added several TODOs and incorporated Alice's 'active member' definition in the voting procedure.

Started defining amendments by pull request - I'm interested in seeing
what you guys think of this. If the doc is on GitHub, we might as well
use GitHub features to edit it down the line!